feat: implement convertBranchNameToPRTitle function and associated tests
- Added the convertBranchNameToPRTitle function to convert commitlint-formatted branch names into readable PR titles.
- Created a new test file for the function, covering various commitlint formats, edge cases, and non-standard formats.
- Removed redundant inline test implementation from open-pr.test.ts to streamline testing structure.
